test(exporters/elasticsearch): wait until all templates are present
Waits until all templates are present in Elasticsearch before starting
the actual test. This is to avoid flakiness when putting templates is
slow, as we have 16 templates to put, and if each takes 1 second, that
adds 16s delay before the first record is exported, which is hard to
write tests against.

To ensure tests always do this, the NonStartableBrokerRule is extracted
from the ExporterIntegrationRule. It's essentially the same, but does
not contain start/stop methods, preventing tests from starting/stopping
the broker manually. Instead, tests must use the abstract class'
startBroker/stopBroker method, where the checks are correctly done.

This is very hacky and very much a temporary measure. The reason I find
this acceptable is that we will be rewriting these tests entirely such
that there is simply no broker, at which point all these things will be
deleted. We committed to doing this before the end of the quarter, and
we don't predict any new ES tests, so fixing the hack (which fixes the
flakiest tests) is definitely valuable.

(cherry picked from commit c9dabf96ae406af12d3524c76a3fb3acab6c3f42)