test(gateway): increase timeouts to avoid instability
The flake occured due to the request having been canceled.

The assertion failure was:

```
Wanted but not invoked:
serverStreamObserver.onCompleted();
-> at io.camunda.zeebe.gateway.api.job.LongPollingActivateJobsTest.shouldUnblockRequestWhenJobsAvailable(LongPollingActivateJobsTest.java:119)
However, there were exactly 4 interactions with this mock:
serverStreamObserver.isCancelled();
-> at io.camunda.zeebe.gateway.impl.job.LongPollingActivateJobsRequest.isCanceled(LongPollingActivateJobsRequest.java:113)
serverStreamObserver.isCancelled();
-> at io.camunda.zeebe.gateway.impl.job.LongPollingActivateJobsRequest.isCanceled(LongPollingActivateJobsRequest.java:113)
serverStreamObserver.isCancelled();
-> at io.camunda.zeebe.gateway.impl.job.LongPollingActivateJobsRequest.isCanceled(LongPollingActivateJobsRequest.java:113)
```

According to the grpc docs:
> Cancellations can be caused by timeouts, explicit cancellation by
> client, network errors, and similar.

A timeout of 5s is pretty long, but might just be enough to cause
problems on a busy CI. Using 1m is much safer. If this times out we
probably see other issues in the CI. So this commit increases these
timeout to a saver value.